fix(exam): 调整默认题目分值及优化批改逻辑代码结构
- 将题目默认分数从5分调整为1分，确保最低分值一致性
- 重构ExamRecordService，使用Lombok简化构造函数和日志注入
- 增加考试多次记录的设计和考试时长字段注释，便于后续功能扩展
- 优化选择题和判断题批改逻辑，提升代码可读性
- 调整部分实体及请求对象中的分数类型为int，增强类型一致性
- 删除Paper相关接口冗余代码及废弃版本，简化接口实现
- 优化QuestionExcelListener中题目导入默认分数逻辑，防止空值异常

+++ exam-system-java/entity/src/main/java/com/yyjy/exam/entity/paper/entity/PaperQuestion.java
@@ -21,7 +21,7 @@ public interface PaperQuestion {
 	@OnDissociate(DissociateAction.DELETE)
 	Questions question();
 	
-	double score();
+	int score();
 	
 	@Nullable
 	LocalDateTime createTime();
